refactor: 상,마 전략 공통 부분 추상화
